Refactor API querystring so the UI can reuse this
