+++ packages/api/src/useCases/cases/fetchCasesAndFilter.ts
@@ -1,9 +1,10 @@
+import type { ApiCaseQuerystring } from "@moj-bichard7/common/types/ApiCaseQuerystring"
 import type { CaseIndexMetadata } from "@moj-bichard7/common/types/Case"
 import type { User } from "@moj-bichard7/common/types/User"
 
 import type DataStoreGateway from "../../services/gateways/interfaces/dataStoreGateway"
 import type { CaseDataForIndexDto } from "../../types/Case"
-import type { CaseIndexQuerystring, Filters, Pagination, SortOrder } from "../../types/CaseIndexQuerystring"
+import type { Filters, Pagination, SortOrder } from "../../types/CaseIndexQuerystring"
 
 import { convertCaseToCaseIndexDto } from "../dto/convertCaseToDto"
 
@@ -41,7 +42,7 @@ const assignNotesAndTriggers = async (
 
 export const fetchCasesAndFilter = async (
   dataStore: DataStoreGateway,
-  query: CaseIndexQuerystring,
+  query: ApiCaseQuerystring,
   user: User
 ): Promise<CaseIndexMetadata> => {
   const pagination: Pagination = { maxPerPage: query.maxPerPage, pageNum: query.pageNum }
